// src/types/Form.tsx
export type Form = {
    website: string;
    email: string;
    companyName: string;
    industry: string;
    companySize: string;
    worker: string;
    meetingTime: string;
    phoneNumber: string;
    contactMethod: string;
  }

export type FormErrors = {
    website: string;
    email: string;
    companyName: string;
    industry: string;
    companySize: string;
    worker: string;
    meetingTime: string;
    phoneNumber: string;
    contactMethod: string;
}